services/ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-multipart==0.0.20
pydantic-settings==2.7.1
# Forensic analysis dependencies
Pillow>=10.4.0
numpy>=1.26.0
python-magic>=0.4.27
PyExifTool>=0.5.6
c2pa-python>=0.6.0
# Fallback EXIF reader (used when ExifTool binary is unavailable)
exifread>=3.0.0
# Phase 2: Enhanced ELA & CNN forensics
opencv-python-headless>=4.9.0
scipy>=1.12.0
# Phase 3: Optical forensics (Moire / wavelet analysis)
PyWavelets>=1.5.0
# Phase 5: Document forensics (PDF analysis)
pypdf>=4.0.0
pyhanko>=0.25.0
cryptography>=42.0.0
# Phase 8: Evidence standardization (PDF report generation)
fpdf2>=2.8.0
# CNN deep learning (optional — graceful degradation if missing)
# torch and photoholmes are installed separately in Dockerfile for layer caching
